// src/app/core/models/Medidas.ts
export interface Medidas {
    id?:                                number;
    redondo:                           number;
    patilla_a_patilla:                 number;
    largo_de_frente:                   number;
    sien_a_sien:                       number;
    oreja_a_oreja_por_encima:          number;
    anchura_del_cuello_superior:      number;
    oreja_a_oreja_por_nacimiento_pelo: number;
    anchura_cuello_inferior:           number;
    createdAt?:                         Date;
    updatedAt?:                         Date;
    user_id:                           number;
}